+++ internal/smartrw/writer.go
@@ -8,31 +8,32 @@ import (
 	"strings"
 )
 
+type Writer struct {
+	Dest string
+	io.WriteCloser
+}
+
 func NewWriter(dest string) (io.WriteCloser, error) {
 	w, err := selectWriter(dest)
 	if err != nil {
 		return nil, err
 	}
 
-	return &writer{
-		dest:        dest,
+	return &Writer{
+		Dest:        dest,
 		WriteCloser: w,
 	}, nil
-}
-
-type writer struct {
-	dest string
-	io.WriteCloser
 }
 
 func selectWriter(dest string) (io.WriteCloser, error) {
 	// empty = write to stdout
 	if len(dest) == 0 {
-		return &StdoutWriter{}, nil
+		return NewStdoutWriter(), nil
 	}
 
 	// handle file:// and protocol-less destinations
 	if strings.HasPrefix(dest, "file://") || !strings.Contains(dest, "://") {
+		dest = strings.TrimPrefix(dest, "file://")
 		w, err := os.Create(dest)
 		if err != nil {
 			return nil, err
